requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pypdf
tiktoken
python-dotenv
